all the files are updated

+++ Datatypes/Datatypes_2.java
@@ -1,7 +1,7 @@
 public class Datatypes_2 {
     
     public static void main(String[] args) {
-    short a =1223232;
+    int a =1223232;
     System.out.println(a);
 
     }
